requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0
aiofiles==24.1.0
pillow==10.4.0
razorpay==1.4.2
python-dotenv==1.0.1
jinja2==3.1.4
weasyprint==62.3
boto3==1.35.0
